% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descfars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{This function plots the map of pointwise car fatalities in given year at a given American state.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numeric vector representing the number of the American state: 1-51 @param years A numeric vector representing the years: 2013-2015}
}
\value{
a map of American state and points of fatalities
}
\description{
This function plots the map of pointwise car fatalities in given year at a given American state.
}
\section{warning when the input characters  cannot be numeric year or numeric statenumber @section warning when the state.number in the in the available data range}{
 1-51. @section warning when the year in the in the available data range: 2013-2015.
}

\examples{
Map with fatalities count in the state of Delaware in 2015 \dontrun{fars_map_state(10,2015)}

}
