% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descfars_functions.R
\name{fars_read}
\alias{fars_read}
\title{This function reads file from .csv input file.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string with the name of the file to read}
}
\value{
An error message if the data does not exist or a data frame with data readed from the csv input file.
}
\description{
This function reads file from .csv input file.
}
\section{warning When the input file does not exist.}{
NA
}

\examples{
\dontrun{fars_read("accident_2015.csv.bz2")}

}
