% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descfars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{This function reads data per input year and returns a tibble data matrix of 2 columns (Month,year).}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A numeric vector representing the years}
}
\value{
a tibble data-matrix of month and year and fatalities in that month
}
\description{
This function reads data per input year and returns a tibble data matrix of 2 columns (Month,year).
}
\section{warning when the input character cannot be numeric year or the year in the in the available data range}{
 2013-2015.
}

\examples{
\dontrun{fars_read_years(2015)}

}
