% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descfars_functions.R
\name{make_filename}
\alias{make_filename}
\title{This function gets year and creates file name based on that input year.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a numeric character representing the year.}
}
\value{
A character vector corresponding to the filenames for the  given year.
}
\description{
This function gets year and creates file name based on that input year.
}
\section{warning When the input character cannot be converted to numeric character}{
NA
}

\examples{
\dontrun{make_filename(2015)}

}
